changed imports/exports and directory referencing
