+++ web-app/client/components/Analytics/AnalyticsStat.js
