+++ web-app/client/components/Analytics/AnalyticsTable.js
