+++ web-app/client/components/Analytics/ReadingHours.js
