+++ web-app/client/components/Analytics/index.js
@@ -1,0 +1,11 @@
+/* Central export for Analytics components */
+
+export { default as Analytics } from './Analytics';
+export { default as AnalyticsFeatured } from './AnalyticsFeatured';
+export { default as AnalyticsStat } from './AnalyticsStat';
+export { default as AnalyticsTable } from './AnalyticsTable';
+export { default as ReadingHours } from './ReadingHours';
+export { default as ReadingSources } from './ReadingSources';
+export { default as ReadingTimeGraphs } from './ReadingTimeGraphs';
+export { default as ReadingTimeline } from './ReadingTimeline';
+export { default as VisContainer } from './VisContainer';
