+++ web-app/client/components/Article/ArticleButtons/ButtonArchiveArticle.js
@@ -1,22 +1,22 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { putArticleStatus } from '../store';
+import { putArticleStatus } from '../../../store';
 import { ButtonIcon } from '.';
 
 /**
  * COMPONENT
  */
 
-const ButtonAddToMyListArticle = props => {
+const ButtonArchiveArticle = props => {
   const { articleId, handleAddToMyListArticle } = props;
 
   return (
     <ButtonIcon
       articleId={articleId}
       handleButtonClick={handleAddToMyListArticle}
-      iconName="newspaper"
-      popupLabel="My List"
-      successMessage="Article added back to reading list"
+      iconName="archive"
+      popupLabel="Archive"
+      successMessage="Article archived"
     />
   );
 };
@@ -24,9 +24,9 @@ const ButtonAddToMyListArticle = props => {
 const mapDispatchToProps = dispatch => {
   return {
     handleAddToMyListArticle: articleId => {
-      dispatch(putArticleStatus(articleId, 'my-list'));
+      dispatch(putArticleStatus(articleId, 'archive'));
     }
   };
 };
 
-export default connect(null, mapDispatchToProps)(ButtonAddToMyListArticle);
+export default connect(null, mapDispatchToProps)(ButtonArchiveArticle);
