+++ web-app/client/components/Article/ArticleButtons/ButtonDeleteArticle.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { deleteArticle } from '../store';
+import { deleteArticle } from '../../../store';
 import { ButtonIcon } from '.';
 
 /**
