+++ web-app/client/components/Article/ArticleButtons/index.js
@@ -1,0 +1,8 @@
+/* Central export for Article Button components */
+
+export { default as ButtonAddArticle } from './ButtonAddArticle';
+export { default as ButtonUserIcon } from './ButtonUserIcon';
+export { default as ButtonArchiveArticle } from './ButtonArchiveArticle';
+export { default as ButtonDeleteArticle } from './ButtonDeleteArticle';
+export { default as ButtonIcon } from './ButtonIcon';
+export { default as ButtonMyListArticle } from './ButtonMyListArticle';
