+++ web-app/client/components/Article/index.js
@@ -1,0 +1,9 @@
+/* Central export for Article components */
+
+export { default as Article } from './Article';
+export { default as ArticleCard } from './ArticleCard';
+export { default as ArticleCardList } from './ArticleCardList';
+export { default as ArticleList } from './ArticleList';
+export { default as ArticleListItem } from './ArticleListItem';
+export { default as FormAddArticle } from './FormAddArticle';
+export * from './ArticleButtons';
