+++ web-app/client/components/ChangeStyle/index.js
@@ -1,0 +1,4 @@
+/* Central export for style change components */
+
+export { default as ButtonChangeStyle } from './ButtonChangeStyle';
+export { default as FormChangeStyle } from './FormChangeStyle';
