+++ web-app/client/components/Navbar/index.js
@@ -1,0 +1,4 @@
+export { default as NavbarTop } from './NavbarTop';
+export { default as NavbarSideDesktop } from './NavbarSideDesktop';
+export { default as NavbarSideMobile } from './NavbarSideMobile';
+export { default as NavbarSideItems } from './NavbarSideItems';
