+++ web-app/client/components/index.js
@@ -4,37 +4,41 @@
  * having to figure out which file they belong to!
  */
 
-export { default as Analytics } from './Analytics';
-export { default as AnalyticsFeatured } from './AnalyticsFeatured';
-export { default as AnalyticsStat } from './AnalyticsStat';
-export { default as AnalyticsTable } from './AnalyticsTable';
-export { default as Article } from './Article';
-export { default as ArticleCard } from './ArticleCard';
-export { default as ArticleCardList } from './ArticleCardList';
-export { default as ArticleList } from './ArticleList';
-export { default as ArticleListItem } from './ArticleListItem';
-export { default as ButtonAddArticle } from './ButtonAddArticle';
-export { default as ButtonUserIcon } from './ButtonUserIcon';
-export { default as ButtonChangeStyle } from './ButtonChangeStyle';
-export { default as ButtonArchiveArticle } from './ButtonArchiveArticle';
-export { default as ButtonDeleteArticle } from './ButtonDeleteArticle';
-export { default as ButtonIcon } from './ButtonIcon';
-export { default as ButtonMyListArticle } from './ButtonMyListArticle';
-export { default as FormAddArticle } from './FormAddArticle';
-export { default as FormChangeStyle } from './FormChangeStyle';
+export * from './Analytics';
+export * from './Article';
+export * from './ChangeStyle';
+export * from './Navbar';
+// export { default as Analytics } from './Analytics';
+// export { default as AnalyticsFeatured } from './AnalyticsFeatured';
+// export { default as AnalyticsStat } from './AnalyticsStat';
+// export { default as AnalyticsTable } from './AnalyticsTable';
+// export { default as Article } from './Article';
+// export { default as ArticleCard } from './ArticleCard';
+// export { default as ArticleCardList } from './ArticleCardList';
+// export { default as ArticleList } from './ArticleList';
+// export { default as ArticleListItem } from './ArticleListItem';
+// export { default as ButtonAddArticle } from './ButtonAddArticle';
+// export { default as ButtonUserIcon } from './ButtonUserIcon';
+// export { default as ButtonChangeStyle } from './ButtonChangeStyle';
+// export { default as ButtonArchiveArticle } from './ButtonArchiveArticle';
+// export { default as ButtonDeleteArticle } from './ButtonDeleteArticle';
+// export { default as ButtonIcon } from './ButtonIcon';
+// export { default as ButtonMyListArticle } from './ButtonMyListArticle';
+// export { default as FormAddArticle } from './FormAddArticle';
+// export { default as FormChangeStyle } from './FormChangeStyle';
 export { default as LandingPage } from './LandingPage';
 export { default as Message } from './Message';
-export { default as NavbarTop } from './NavbarTop';
-export { default as NavbarSideDesktop } from './NavbarSideDesktop';
-export { default as NavbarSideMobile } from './NavbarSideMobile';
-export { default as NavbarSideItems } from './NavbarSideItems';
+// export { default as NavbarTop } from './NavbarTop';
+// export { default as NavbarSideDesktop } from './NavbarSideDesktop';
+// export { default as NavbarSideMobile } from './NavbarSideMobile';
+// export { default as NavbarSideItems } from './NavbarSideItems';
 export { default as PageContainer } from './PageContainer';
-export { default as ReadingHours } from './ReadingHours';
-export { default as ReadingSources } from './ReadingSources';
-export { default as ReadingTimeGraphs } from './ReadingTimeGraphs';
-export { default as ReadingTimeline } from './ReadingTimeline';
+// export { default as ReadingHours } from './ReadingHours';
+// export { default as ReadingSources } from './ReadingSources';
+// export { default as ReadingTimeGraphs } from './ReadingTimeGraphs';
+// export { default as ReadingTimeline } from './ReadingTimeline';
 export { default as UserHome } from './UserHome';
-export { default as VisContainer } from './VisContainer';
+// export { default as VisContainer } from './VisContainer';
 export { default as Home } from './Home';
 export { default as LoginPage } from './LoginPage';
 export { default as SignupPage } from './SignupPage';
